+++ plugins/lightningcss.ts
@@ -49,7 +49,10 @@ export const defaults: Options = {
 /** A plugin to load all CSS files and process them using parcelCSS */
 export default function (userOptions?: DeepPartial<Options>) {
   return (site: Site) => {
-    const options = merge(defaults, userOptions);
+    const options = merge<Options>(
+      { ...defaults, includes: site.options.includes },
+      userOptions,
+    );
 
     site.loadAssets(options.extensions);