+++ tests/lightningcss.test.ts
@@ -1,12 +1,14 @@
 import { assertSiteSnapshot, build, getSite } from "./utils.ts";
 import lightningcss from "../plugins/lightningcss.ts";
 
-Deno.test("lightningcss plugin", async (t) => {
+Deno.test("lightningcss plugin (only transform)", async (t) => {
   const site = getSite({
     src: "lightningcss",
   });
 
-  site.use(lightningcss());
+  site.use(lightningcss({
+    includes: false,
+  }));
 
   await build(site);
   await assertSiteSnapshot(t, site);
@@ -17,9 +19,7 @@ Deno.test("lightningcss plugin (bundle mode)", async (t) => {
     src: "lightningcss",
   });
 
-  site.use(lightningcss({
-    includes: "_includes",
-  }));
+  site.use(lightningcss());
 
   await build(site);
   await assertSiteSnapshot(t, site);