+++ CHANGELOG.md
@@ -13,6 +13,10 @@ Any BREAKING CHANGE between minor versions will be documented here in upper case
 - TOML Plugin [#432]
 - JSON Plugin supports `.jsonc` files [#433]
 
+### Changed
+- `lightningcss` plugin bundles the CSS code by default.
+  Set the option `includes: false` to only transform the code.
+
 ## [1.17.5] - 2023-06-08
 ### Fixed
 - `Site.copy` now works as expected when given a path with a trailing slash. [#426]